open close into single cog
